+++ didums-hk2/src/test/java/com/github/bordertech/didums/hk2/TestQualifierImpl.java
@@ -1,0 +1,10 @@
+package com.github.bordertech.didums.hk2;
+
+import org.glassfish.hk2.api.AnnotationLiteral;
+
+/**
+ * Instance of the test qualifier annotation to use on binding test.
+ */
+public class TestQualifierImpl extends AnnotationLiteral<TestQualifier> implements TestQualifier {
+
+}