+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Latest qa-parent
 * Switch from travis-ci to GitHub Actions #23
+* Update HK2 dependency to the latest version that supports javax.inject API #24
 
 ## 1.0.4